feat(kibana-proxy)
added build proccess (copy) to proxy and added a run.sh script that runs both kibana and proxy.
also changed proxy to redirect with the same domain and protocol the user accessed the proxy.

CFY-2943
