--- proxy/conf.json
@@ -1,5 +1,0 @@
-{
-    "port": 5602,
-    "kibana_ip": "localhost",
-    "kibana_port": 5601
-}
