+++ src/proxy/config/conf.json
@@ -1,0 +1,3 @@
+{
+    "port": 5602
+}
